consolidate all CustomDefinitions under one folder
