+++ SolastaCommunityExpansion/Classes/Tinkerer/Subclasses/ScoutSentinelSubClass.cs
@@ -2,7 +2,7 @@
 using System.Runtime.Serialization;
 using SolastaCommunityExpansion.Builders;
 using SolastaCommunityExpansion.Builders.Features;
-using SolastaCommunityExpansion.CustomFeatureDefinitions;
+using SolastaCommunityExpansion.CustomDefinitions;
 using SolastaModApi;
 using SolastaModApi.Extensions;
 using SolastaModApi.Infrastructure;
