+++ SolastaCommunityExpansion/Classes/Tinkerer/TinkererClass.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using SolastaCommunityExpansion.Builders;
 using SolastaCommunityExpansion.Builders.Features;
 using SolastaCommunityExpansion.Classes.Tinkerer.Subclasses;
-using SolastaCommunityExpansion.CustomFeatureDefinitions;
+using SolastaCommunityExpansion.CustomDefinitions;
 using SolastaModApi;
 using static CharacterClassDefinition;
 using static SolastaModApi.DatabaseHelper;
