+++ SolastaCommunityExpansion/CustomDefinitions/FeatureDefinitionCustomCode.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SolastaCommunityExpansion.Builders.Features;
 
-namespace SolastaCommunityExpansion.CustomFeatureDefinitions
+namespace SolastaCommunityExpansion.CustomDefinitions
 {
     /**
      * This adds the ability to do fully custom code when a feature is granted. Where possible you should use more targeted features,
