+++ SolastaCommunityExpansion/CustomDefinitions/IClassHoldingFeature.cs
@@ -1,4 +1,4 @@
-﻿namespace SolastaCommunityExpansion.CustomFeatureDefinitions
+﻿namespace SolastaCommunityExpansion.CustomDefinitions
 {
     /// <summary>
     /// Implement on an IAdditionalDamageProvider feature to allow its damage to scale with class level (using DiceByRankTable), even if it isn't directly added to the class (via AddFeatureAtLevel).
