+++ SolastaCommunityExpansion/CustomDefinitions/INotifyConditionRemoval.cs
@@ -1,4 +1,4 @@
-﻿namespace SolastaCommunityExpansion.CustomFeatureDefinitions
+﻿namespace SolastaCommunityExpansion.CustomDefinitions
 {
     /// <summary>
     /// Implement on a ConditionDefinition to be notified when a condition is removed, or when a creature is about to die with a condition.
