+++ SolastaCommunityExpansion/CustomDefinitions/IStartOfTurnRecharge.cs
@@ -1,4 +1,4 @@
-﻿namespace SolastaCommunityExpansion.CustomFeatureDefinitions
+﻿namespace SolastaCommunityExpansion.CustomDefinitions
 {
     /// <summary>
     /// Implement on a FeatureDefinitionPower to allow it to recharge at the start of your turn.
