+++ SolastaCommunityExpansion/Level20/Features/IndomitableMightBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using SolastaCommunityExpansion.Builders;
 using SolastaCommunityExpansion.Builders.Features;
-using SolastaCommunityExpansion.CustomFeatureDefinitions;
+using SolastaCommunityExpansion.CustomDefinitions;
 using SolastaModApi.Extensions;
 
 namespace SolastaCommunityExpansion.Level20.Features
