+++ SolastaCommunityExpansion/Models/CustomFeaturesContext.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using SolastaCommunityExpansion.CustomFeatureDefinitions;
+using SolastaCommunityExpansion.CustomDefinitions;
 using SolastaModApi.Extensions;
 using UnityEngine;
