+++ SolastaCommunityExpansion/Patches/CustomFeatures/StartOfTurnRecharge/RulesetCharacterPatcher.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using HarmonyLib;
-using SolastaCommunityExpansion.CustomFeatureDefinitions;
+using SolastaCommunityExpansion.CustomDefinitions;
 
 namespace SolastaCommunityExpansion.Patches.CustomFeatures.StartOfTurnRecharge
 {
